requirements.txt
pandas
numpy
matplotlib
geopandas
shapely
sqlalchemy
mysql-connector-python
pymysql